+++ packages/harpparts/src/apparatus/index.ts
@@ -4,6 +4,10 @@ export {
   NATURAL_MINOR_APPARATUS,
   WILDE_TUNED_APPARATUS,
   POWER_BENDER_APPARATUS,
+  POWER_DRAW_APPARATUS,
+  POWER_DRAW_HALF_VALVED_APPARATUS,
+  WOOZLE_MINOR_APPARATUS,
+  WOOZLE_MINOR_HALF_VALVED_APPARATUS,
 } from './constants'
 export { ApparatusIds } from './types'
 export type { Apparatus } from './types'