+++ packages/harpstrata/src/get-harp-strata/utils/get-pitch-matrix/get-pitch-matrix.test.ts
@@ -5,6 +5,8 @@ import { EXAMPLE_PITCH_MATRICES } from '../../testResources'
 import { getPitchMatrix } from './index'
 
 const c = getPitch(PitchIds.C)
+const b = getPitch(PitchIds.B)
+const aFlat = getPitch(PitchIds.Ab)
 
 test('getPitchMatrix function maps a simple 2d array of 0s to the input key pitch of C', () => {
   const expectedArray = [[c], [c]]
@@ -31,4 +33,18 @@ test('getPitchMatrix maps a major diatonic halfstepmatrix in to a major diatonic
   )
 
   expect(actualArray).toStrictEqual(MAJOR_DIATONIC_F_HARMONICA)
+})
+
+test('getPitchMatrix function maps a simple 2d array of -1s to Bs for a C harmonica', () => {
+  const expectedArray = [[b], [b]]
+  const actualArray = getPitchMatrix([[-1], [-1]], PitchIds.C)
+
+  expect(actualArray).toStrictEqual(expectedArray)
+})
+
+test('getPitchMatrix function maps a simple 2d array of -13s to Abs for an A harmonica', () => {
+  const expectedArray = [[aFlat], [aFlat]]
+  const actualArray = getPitchMatrix([[-13], [-13]], PitchIds.A)
+
+  expect(actualArray).toStrictEqual(expectedArray)
 })