+++ packages/harpstrata/CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to ~~[Semantic Versioning](https://semver.org/spec/v2.0
 
 ## [Unreleased](https://github.com/js-jslog/harpguru/compare/v5.0.0...HEAD) - yyyy-mm-dd
 
+### Fixed
+
+- MINOR: Correctly handle negative halfstep index's in processing of apparatus from harpparts module
+
 ## [v8.0.0](https://github.com/js-jslog/harpguru/releases/tag/v5.0.0) - 2020-12-04
 
 ### Changed